Viscacha 0.8.1.2

+++ addreply.php
@@ -74,18 +74,19 @@ if ($config['tpcallow'] == 1 && $my->p['attachments'] == 1) {
 	$p_upload = 1;
 }
 
+$validDigest = array(-1, 1, 2, 3, 9);
 $standard_data = array(
-			'name' => '',
-			'email' => '',
-			'guest' => iif($my->vlogin, 0, 1),
-			'comment' => '',
-			'dosmileys' => 1,
-			'dowords' => 1,
-			'digest' => -1,
-			'topic' => $lang->phrase('reply_prefix').$info['topic'],
-			'human' => false,
-			'id' => $id
-		);
+	'name' => '',
+	'email' => '',
+	'guest' => iif($my->vlogin, 0, 1),
+	'comment' => '',
+	'dosmileys' => 1,
+	'dowords' => 1,
+	'digest' => 0,
+	'topic' => $lang->phrase('reply_prefix').$info['topic'],
+	'human' => false,
+	'id' => $id
+);
 
 ($code = $plugins->load('addreply_start')) ? eval($code) : null;
 
@@ -223,29 +224,26 @@ if ($_GET['action'] == "save") {
 		// Set uploads to correct reply
 		$db->query("UPDATE {$db->pre}uploads SET tid = '{$redirect}' WHERE mid = '{$pid}' AND topic_id = '{$id}' AND tid = '0'");
 
-		// Update, insert, delete notifications
-		if ($my->vlogin) {
-			$result = $db->query("SELECT id, type FROM {$db->pre}abos WHERE mid = '{$my->id}' AND tid = '{$id}'");
+		// Update, insert notifications
+		if ($my->vlogin && in_array($digest, $validDigest)) {
 			switch ($digest) {
 				case 1:  $type = '';  break;
 				case 2:  $type = 'd'; break;
 				case 3:  $type = 'w'; break;
-				default: $type = null; break;
+				case 9:  $type = 'f'; break;
+				case -1: $type = null; break;
 			}
-
-			if ($db->num_rows($result) > 0) {
-				$row = $db->fetch_assoc($result);
-				if ($type === null) { // Lösche Abo
-					$db->query("DELETE FROM {$db->pre}abos WHERE id = '{$row['id']}'");
-				}
-				elseif ($row['type'] != $type) { // Aktualisiere Abo, wenn veränderter Typ
-					$db->query("UPDATE {$db->pre}abos SET type = '{$type}' WHERE id = '{$row['id']}'");
-				}
+			
+			$result = $db->query("SELECT id, type FROM {$db->pre}abos WHERE mid = '{$my->id}' AND tid = '{$id}'");
+			$row = $db->fetch_assoc($result);
+			if ($row && $type === null) { // Lösche Abo
+				$db->query("DELETE FROM {$db->pre}abos WHERE id = '{$row['id']}'");
 			}
-			else {
-				if ($type !== null) { // Füge Abo hinzu
-					$db->query("INSERT INTO {$db->pre}abos (mid, tid, type) VALUES ('{$my->id}', '{$id}', '{$type}')");
-				}
+			else if ($row && $type != $row['type']) { // Aktualisiere Abo, wenn veränderter Typ
+				$db->query("UPDATE {$db->pre}abos SET type = '{$type}' WHERE id = '{$row['id']}'");
+			}
+			else if (!$row && $type !== null) { // Füge Abo hinzu
+				$db->query("INSERT INTO {$db->pre}abos (mid, tid, type) VALUES ('{$my->id}', '{$id}', '{$type}')");
 			}
 		}
 
@@ -321,6 +319,7 @@ else {
 
 	($code = $plugins->load('addreply_form_start')) ? eval($code) : null;
 
+	$data['formatted_comment'] = '';
 	if (is_hash($fid)) {
 		$data = $gpc->unescape(import_error_data($fid));
 		if ($id != $data['id']) {
@@ -391,7 +390,7 @@ else {
 		$captcha = null;
 	}
 
-	if ($my->vlogin && $data['digest'] == -1) {
+	if ($my->vlogin && !in_array($data['digest'], $validDigest)) {
 		$result = $db->query("SELECT type FROM {$db->pre}abos WHERE mid = '{$my->id}' AND tid = '{$id}'");
 		if ($db->num_rows($result) > 0) {
 			$temp = $db->fetch_assoc($result);
@@ -399,11 +398,12 @@ else {
 				case '':	$data['digest'] = 1; break;
 				case 'd':	$data['digest'] = 2; break;
 				case 'w':	$data['digest'] = 3; break;
-				default:	$data['digest'] = 0; // Favoriten = Keine Benachrichtigung
+				case 'f':	$data['digest'] = 9; break;
+				default:	$data['digest'] = -1;
 			}
 		}
 		else {
-			$data['digest'] = 0;
+			$data['digest'] = -1;
 		}
 	}
